Major Backend Overhaul for Improved Structure and Scalability
This is a large makeover of our current code to essentially declutter files and make room for future work.

Here are the changes:
new folder structure --- there are now four new folders for our backend which all have filetypes that will handle different scenarios.
controllers : this will handle routing logic and connect the frontend with backend services. We dont use this yet because itll mainly be for when we implement databases to store whole games and chatlogs etc etc
models: these define data structures and ensure consistency across the app
socket: this manages all our socketio stuff
services: this is for the business logic and handling of core app operations.

I also fixed our chatroom issue of having it all sent on the same side. The issue was that our clients kept thinking we were all the same person and thats why it appeared as though it was all sent on the user sending side even though we tried to view it from other users perspective it still showed we sent it.

I fixed this by overhauling how we handled usernames in general, now we handle it with a player object which is held inside the lobby object. Chatroom appears to work fine now.

In the future I want to implement more documentation and detailed comments across all new files so everyone knows whats going on and how they can start their work. Also front end needs to be worked on